+++ seeqret/storage/storage.py
@@ -11,6 +11,9 @@ class Storage:
     def add_user(self, user: User):
         raise NotImplementedError
 
+    def remove_user(self, username: str):
+        raise NotImplementedError   # pragma: no cover
+
     def fetch_user(self, username: str):
         raise NotImplementedError   # pragma: no cover